#9 #24 refactor: accommodation과 room 패키지 분리
