+++ src/main/java/com/backoffice/upjuyanolja/domain/accommodation/entity/Accommodation.java
@@ -1,5 +1,6 @@
 package com.backoffice.upjuyanolja.domain.accommodation.entity;
 
+import com.backoffice.upjuyanolja.domain.room.entity.Room;
 import com.backoffice.upjuyanolja.global.common.entity.BaseTime;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
