--- src/main/java/com/backoffice/upjuyanolja/domain/accommodation/repository/AccommodationImageRepository.java
@@ -1,8 +1,0 @@
-package com.backoffice.upjuyanolja.domain.accommodation.repository;
-
-import com.backoffice.upjuyanolja.domain.accommodation.entity.AccommodationImage;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationImageRepository extends JpaRepository<AccommodationImage, Long> {
-
-}
