--- src/main/java/com/backoffice/upjuyanolja/domain/accommodation/repository/RoomImageRepository.java
@@ -1,8 +1,0 @@
-package com.backoffice.upjuyanolja.domain.accommodation.repository;
-
-import com.backoffice.upjuyanolja.domain.accommodation.entity.RoomImage;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoomImageRepository extends JpaRepository<RoomImage, Long> {
-
-}
