--- src/main/java/com/backoffice/upjuyanolja/domain/accommodation/repository/RoomRepository.java
@@ -1,8 +1,0 @@
-package com.backoffice.upjuyanolja.domain.accommodation.repository;
-
-import com.backoffice.upjuyanolja.domain.accommodation.entity.Room;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoomRepository extends JpaRepository<Room, Long> {
-
-}
