+++ src/main/java/com/backoffice/upjuyanolja/domain/openapi/service/OpenApiService.java
@@ -5,16 +5,16 @@ import com.backoffice.upjuyanolja.domain.accommodation.entity.AccommodationImage
 import com.backoffice.upjuyanolja.domain.accommodation.entity.AccommodationOption;
 import com.backoffice.upjuyanolja.domain.accommodation.entity.Address;
 import com.backoffice.upjuyanolja.domain.accommodation.entity.Category;
-import com.backoffice.upjuyanolja.domain.accommodation.entity.Room;
-import com.backoffice.upjuyanolja.domain.accommodation.entity.RoomImage;
-import com.backoffice.upjuyanolja.domain.accommodation.entity.RoomOption;
-import com.backoffice.upjuyanolja.domain.accommodation.entity.RoomPrice;
 import com.backoffice.upjuyanolja.domain.accommodation.repository.AccommodationImageRepository;
 import com.backoffice.upjuyanolja.domain.accommodation.repository.AccommodationRepository;
-import com.backoffice.upjuyanolja.domain.accommodation.repository.RoomRepository;
 import com.backoffice.upjuyanolja.domain.openapi.exception.InvalidDataException;
 import com.backoffice.upjuyanolja.domain.openapi.exception.OpenApiException;
-import com.backoffice.upjuyanolja.domain.accommodation.repository.RoomImageRepository;
+import com.backoffice.upjuyanolja.domain.room.entity.Room;
+import com.backoffice.upjuyanolja.domain.room.entity.RoomImage;
+import com.backoffice.upjuyanolja.domain.room.entity.RoomOption;
+import com.backoffice.upjuyanolja.domain.room.entity.RoomPrice;
+import com.backoffice.upjuyanolja.domain.room.repository.RoomImageRepository;
+import com.backoffice.upjuyanolja.domain.room.repository.RoomRepository;
 import jakarta.annotation.PostConstruct;
 import java.net.URI;
 import java.net.URLEncoder;
